New code integrated but tests are failing
