+++ codegen/templates/mat.rs.tera
@@ -72,7 +72,7 @@ use crate::{
         Mat3A,
     {% endif %}
 {% elif dim == 3 %}
-    EulerRot,
+    EulerRot, euler::FromEuler,
     {{ mat2_t }}, {{ mat4_t }}, {{ quat_t }}, {{ vec2_t }}, {{ col_t }},
     {% if is_align %}
         {{ mat3_t }}, {{ vec3_t }},
@@ -80,7 +80,7 @@ use crate::{
         Mat3A, Vec3A,
     {% endif %}
 {% elif dim == 4 %}
-    EulerRot,
+    EulerRot, euler::FromEuler,
     {{ mat3_t }}, {{ quat_t }}, {{ vec3_t }}, {{ col_t }},
     {% if scalar_t == "f32" %}
         Mat3A, Vec3A,
@@ -521,8 +521,7 @@ impl {{ self_t }} {
     #[inline]
     #[must_use]
     pub fn from_euler(order: EulerRot, a: {{ scalar_t }}, b: {{ scalar_t }}, c: {{ scalar_t }}) -> Self {
-        let quat = {{ quat_t }}::from_euler(order, a, b, c);
-        Self::from_quat(quat)
+        Self::from_euler_angles(order, a, b, c)
     }
 
     /// Creates a 3D rotation matrix from `angle` (in radians) around the x axis.
@@ -853,8 +852,7 @@ impl {{ self_t }} {
     #[inline]
     #[must_use]
     pub fn from_euler(order: EulerRot, a: {{ scalar_t }}, b: {{ scalar_t }}, c: {{ scalar_t }}) -> Self {
-        let quat = {{ quat_t }}::from_euler(order, a, b, c);
-        Self::from_quat(quat)
+        Self::from_euler_angles(order, a, b, c)
     }
 
     /// Creates an affine transformation matrix containing a 3D rotation around the x axis of
