+++ src/f32/scalar/quat.rs
@@ -1,8 +1,7 @@
 // Generated from quat.rs.tera template. Edit the template, not the generated file.
 
 use crate::{
-    // euler::{EulerFromQuaternion, EulerRot, EulerToQuaternion},
-    euler2::{QuatToEuler, QuatFromEuler, EulerRot},
+    euler::{EulerRot, FromEuler, ToEuler},
     f32::math,
     DQuat, Mat3, Mat3A, Mat4, Vec2, Vec3, Vec3A, Vec4,
 };
@@ -185,7 +184,6 @@ impl Quat {
     #[must_use]
     pub fn from_euler(euler: EulerRot, a: f32, b: f32, c: f32) -> Self {
         Self::from_euler_angles(euler, a, b, c)
-        // euler.new_quat(a, b, c)
     }
 
     /// From the columns of a 3x3 rotation matrix.
@@ -396,9 +394,8 @@ impl Quat {
     /// Returns the rotation angles for the given euler rotation sequence.
     #[inline]
     #[must_use]
-    pub fn to_euler(self, euler: EulerRot) -> (f32, f32, f32) {
-        // euler.convert_quat(self)
-        self.extract_angles(euler)
+    pub fn to_euler(self, order: EulerRot) -> (f32, f32, f32) {
+        self.to_euler_angles(order)
     }
 
     /// `[x, y, z, w]`
